Extra test case and methods are added for user create api

+++ Mutual_Methods/src/main/java/imp/methods/DeleteRequestImp.java
@@ -3,13 +3,19 @@ package imp.methods;
 import com.thoughtworks.gauge.Step;
 import exceptions.RequestNotDefined;
 import helper.methods.DeleteHelper;
+import utils.Utils;
 
 public class DeleteRequestImp extends DeleteHelper {
 
 
     @Step({"Delete request", "Delete isteği gönder"})
     public void deleteRequests() throws RequestNotDefined {
-        deleteRequest();
+        String isDelete = String.valueOf(Utils.getFromStoreData("deleteAfterTest"));
+        if (!isDelete.equals("No")) {
+
+            deleteRequest();
+        }
+        System.out.println("Deletion Result:"+isDelete);
     }
 
     @Step({"Delete request <url>", "Delete isteği gönder <url>"})
